move src/state_content.ts into src/content
