+++ src/content/controller_content.ts
@@ -1,5 +1,5 @@
 import { isTextEditable } from "@src/dom"
-import { contentState, ModeName } from "@src/state_content"
+import { contentState, ModeName } from "@src/content/state_content"
 import { repeat } from "@src/.excmds_background.generated"
 import Logger from "@src/lib/logging"
 import * as messaging from "@src/lib/messaging"
