+++ src/excmds.ts
@@ -95,7 +95,7 @@ import * as Messaging from "@src/lib/messaging"
 import { browserBg, activeTabId, activeTabContainerId, openInNewTab, openInNewWindow } from "./lib/webext"
 import * as Container from "./lib/containers"
 import state from "./state"
-import { contentState, ModeName } from "./state_content"
+import { contentState, ModeName } from "@src/content/state_content"
 import * as UrlUtil from "./url_util"
 import * as config from "@src/lib/config"
 import * as aliases from "@src/lib/aliases"
