+++ src/parsers/gobblemode.ts
@@ -1,4 +1,4 @@
-import { contentState } from "../state_content"
+import { contentState } from "@src/content/state_content"
 import { isSimpleKey } from "../keyseq"
 
 /** Simple container for the gobble state. */
